requirements.txt
pymupdf
pdfplumber
pytesseract
Pillow
spacy
sqlalchemy
pymysql
pypdf
